requirements.txt
Flask
requests
python-dotenv
boto3
Werkzeug
runpod==1.6.0
celery